refactor: combine constants files
