--- apis/nucleus/src/components/listbox/components/ListBoxRowColumn/helpers/constants.js
@@ -1,4 +1,0 @@
-export const frequencyTextNone = '-';
-export const barPadPx = 4;
-export const barBorderWidthPx = 1;
-export const barWithCheckboxLeftPadPx = 29;
