+++ apis/nucleus/src/components/listbox/components/ListBoxRowColumn/helpers/operations.js
@@ -1,4 +1,4 @@
-import { frequencyTextNone } from './constants';
+import { frequencyTextNone } from '../../../constants';
 
 export const joinClassNames = (namesArray) =>
   namesArray
